create assigne data for the student enrollement table

+++ lms_system/src/main/java/lk/ijse/lms_system/entity/StudentEnrollment.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class StudentEnrollment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer enrollmentId;
+    private Long enrollmentId;
     @ManyToOne
     private Student student;
     @ManyToOne
